renaming file and variables
