+++ lib/helpful.js
